Make all markdown URLs absolute instead of relative
Wiki.js v2 does not work correctly with relative URLs in markdown
(see https://github.com/Requarks/wiki/issues/1201).

As a workaround for this issue, we can convert all relative URLs to
absolute ones. This is not a portable solution (e.g. if we ever serve
the wiki content from a subdirectory), but it should enable us to
switch seamlessly between Wiki.js v1 and v2 as we decide which version
we want to run.